+++ tests/Browser/HomePageTest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Tests\Browser;
+
+use Laravel\Dusk\Browser;
+use Tests\DuskTestCase;
+
+class HomePageTest extends DuskTestCase
+{
+    /**
+     * 尚未有任何條目時，條目 home 顯示提示訊息
+     *
+     * @group homePage
+     */
+    public function testHomeEmpty()
+    {
+        $this->browse(function (Browser $browser) {
+            $browser->visit('/')
+                    ->assertPathIs('/read/home')
+                    ->assertSee('Welcome to Wiki.')
+                    ->assertSeeLink('Start to write something.');
+        });
+    }
+}
